third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/properties/tab-size-expected.txt

This is a testharness.js-based test.
[FAIL] Can set 'tab-size' to a number: 3.14
  assert_equals: expected "CSSUnitValue" but got "CSSMathSum"
Harness: the test ran to completion.
